+++ frontend/src/components/TripList/SearchBar/DestinationList/index.jsx
