+++ frontend/src/components/TripList/SearchBar/Searchbar/index.jsx
